refactor(plugin-sdk): extract replay dedupe contracts
