+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-7c0130372fa6a4e937ffd268a901f735a68a6595b70e4bd9bb8c70059e15ca53  plugin-sdk-api-baseline.json
-b741bcfe31fa7f5a5063e23540713a26d03429f8f3eaf01566cdd9fe0e73af64  plugin-sdk-api-baseline.jsonl
+e59682fe6e11b4bd6e47d585105b75f417ccc774e7085569336fffd478200be2  plugin-sdk-api-baseline.json
+4db490b8a21d1b4ce20f3e4d2f053b034d3ae039c79b2dbba995d13a5c4b0150  plugin-sdk-api-baseline.jsonl
